% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_parXtreem}
\alias{clip_parXtreem}
\title{clip parXtreem table by date}
\usage{
clip_parXtreem(x, startstop = NA)
}
\arguments{
\item{x}{tibble, parXtreem}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
clip parXtreem table by date
}
