% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_tiltometer}
\alias{clip_tiltometer}
\title{clip tiltometer table by date}
\usage{
clip_tiltometer(x, startstop = NA)
}
\arguments{
\item{x}{tibble, tiltometer}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
clip tiltometer table by date
}
