% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_model}
\alias{create_model}
\title{read in model dataset, generate model}
\usage{
create_model(modeldata = create_model_data(), output = NA)
}
\arguments{
\item{modeldata}{tibble, tibble of model data with columns DateTime, sensor, satellite}

\item{output}{character, output filename for RDS object to be saved as}
}
\value{
R object of model
}
\description{
read in model dataset, generate model
}
