% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_model_data}
\alias{create_model_data}
\title{read in sensor and satellite data, generate dataset to create model off of}
\usage{
create_model_data(
  sensordata = example_filename_sensor(),
  satellitedata = example_filename_sat(),
  param = "temp",
  output = NA
)
}
\arguments{
\item{sensordata}{character, the name of the file with sensor data}

\item{satellitedata}{character, the name of the file with satellite data}

\item{param}{character, the type of data being measured, default "temp"}

\item{output}{character, name for outputted dataset to be written as}
}
\value{
dataframe for model creation
}
\description{
read in sensor and satellite data, generate dataset to create model off of
}
