% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_model_plot}
\alias{draw_model_plot}
\title{Plot sensor vs. satellite data}
\usage{
draw_model_plot(
  x = predict_data(),
  param = "Temp",
  main = "Sensor, Satellite, & Predicted Temperature Allen Island",
  xlabel = "Date",
  ylabel = "Temperature (Degrees C)",
  ...
)
}
\arguments{
\item{x}{tibble of sensor & satellite model data}

\item{param}{chatacter, parameter being measured}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot sensor vs. satellite data
}
