% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_par_plot}
\alias{draw_par_plot}
\title{Plot PAR data as line plot}
\usage{
draw_par_plot(
  x = read_parXtreem(),
  main = "PAR",
  xlabel = "Date",
  ylabel = "PAR (umols / m squared / second)",
  facet = NULL,
  ...
)
}
\arguments{
\item{x}{tibble of parXtreem data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{facet}{character, name of the column to facet upon (like "Site") or NULL to skip}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot PAR data as line plot
}
