% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_satsensor_plot}
\alias{draw_satsensor_plot}
\title{Plot sensor vs. satellite data}
\usage{
draw_satsensor_plot(
  x = predict_data(),
  main = "Sensor vs. Satellite Data",
  xlabel = "Sensor",
  ylabel = "Satellite",
  ...
)
}
\arguments{
\item{x}{tibble of sensor & satellite model data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot sensor vs. satellite data
}
