% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{example_filename_par}
\alias{example_filename_par}
\title{retreive example type parXtreem file name}
\usage{
example_filename_par()
}
\value{
filename
}
\description{
retreive example type parXtreem file name
}
