% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{example_filename_sensor}
\alias{example_filename_sensor}
\title{retreive example type sensor file name}
\usage{
example_filename_sensor()
}
\value{
filename
}
\description{
retreive example type sensor file name
}
