% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{example_filename_temp}
\alias{example_filename_temp}
\title{retrieve example type hobotemp file name}
\usage{
example_filename_temp()
}
\value{
filename
}
\description{
retrieve example type hobotemp file name
}
