% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{example_filename_tilt}
\alias{example_filename_tilt}
\title{retreive example type tiltometer file name}
\usage{
example_filename_tilt()
}
\value{
filename
}
\description{
retreive example type tiltometer file name
}
