% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{predict_data}
\alias{predict_data}
\title{read in model, predict days of interest}
\usage{
predict_data(
  model = create_model(),
  modeldata = create_model_data(),
  output = NA
)
}
\arguments{
\item{model}{model, generated by create_model}

\item{modeldata}{tibble, tibble of model data with columns DateTime, sensor, satellite}

\item{output}{character, output filename for final}
}
\value{
dataframe of outputted model values alongside original input values
}
\description{
read in model, predict days of interest
}
