% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_parXtreem}
\alias{read_parXtreem}
\title{read parXtreem data file}
\usage{
read_parXtreem(
  filename = example_filename_par(),
  site = NA,
  output = NA,
  clipped = c("auto", "user", "none")[1],
  startstop = NA
)
}
\arguments{
\item{filename}{character, the name of the file}

\item{site}{character, site being read in}

\item{output}{character, the name for the outputted QAQC file}

\item{clipped}{character, if auto, removed partial start/end days. if user, uses supplied startstop days. if none, does no date trimming}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
read parXtreem data file
}
