% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible-package.R
\docType{package}
\name{sensible}
\alias{sensible-package}
\alias{sensible}
\alias{sensor}
\alias{processing}
\alias{and}
\alias{modeling}
\title{sensible}
\description{
This will process your hobo temperature data.
}
