% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{summarize_parXtreem}
\alias{summarize_parXtreem}
\title{print out summary of parXtreem data}
\usage{
summarize_parXtreem(x = read_parXtreem())
}
\arguments{
\item{x}{tibble, tibble of PAR data}
}
\value{
tibble
}
\description{
print out summary of parXtreem data
}
