% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{summarize_tiltometer}
\alias{summarize_tiltometer}
\title{print out summary of tiltmeter data}
\usage{
summarize_tiltometer(x = read_tiltometer())
}
\arguments{
\item{x}{tibble, tibble of tiltmeter data}
}
\value{
tibble
}
\description{
print out summary of tiltmeter data
}
