% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{tiltometer_rose}
\alias{tiltometer_rose}
\title{plot windrose for currents}
\usage{
tiltometer_rose(
  x = read_tiltometer(),
  legend_title = "Current Speed",
  title = NA,
  speed.cuts = NA,
  flip = TRUE,
  ...
)
}
\arguments{
\item{x}{tibble of tiltometer data}

\item{legend_title}{title of legend}

\item{title}{character, or NA. Title of the plot, or NA to skip}

\item{speed.cuts}{character, numeric, or NA.
\itemize{
\item NA - the default, auto compute 5 quantiles
\item character - "quantile-3" or "quantile-4", computer 3 or 4 quantiles
\item numeric - explicit cut points in a vector
}}

\item{flip}{plotting water, not wind, default TRUE}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot
}
\description{
this is a wrapper around \code{\link[clifro]{windrose}}
}
